+++ CHANGELOG.md
@@ -9,6 +9,12 @@ edit: false
 ## :nut_and_bolt: Other
 ## :lock: Security Fixes
 -->
+
+# v0.6.2 / 2019-10-17
+## :tada: Enhancements
+- [#186](https://github.com/linewalks/MDwalks-UI/issues/186'): Modal 에 description option 추가 ([**@toyjhlee**](https://github.com/toyjhlee))
+- [#181](https://github.com/linewalks/MDwalks-UI/issues/181'): Button, Modal 스타일 조정 ([**@toyjhlee**](https://github.com/toyjhlee))
+
 # v0.6.1 / 2019-10-11
 ## :tada: Enhancements
 - [#178](https://github.com/linewalks/MDwalks-UI/issues/178'): Modal component 생성 ([**@toyjhlee**](https://github.com/toyjhlee))